// src/types/weather/params.d.ts
type TemperatureUnit = "celsius" | "fahrenheit";

type Timeformat = "unixtime" | "iso8601";

type WindSpeedUnit = "kmh" | "ms" | "mph" | "kn";

type PrecipitationUnit = "mm" | "inch";

type CellSelection = "" | "sea" | "nearest";

type TemporalResolution = "" | "hourly_3" | "hourly_6" | "native";

type Timezone =
  //#region Timezones
  | "auto"
  | "UTC"
  | "GMT"
  | "America/Anchorage"
  | "America/Los_Angeles"
  | "America/Denver"
  | "America/Chicago"
  | "America/New_York"
  | "America/Sao_Paulo"
  | "Europe/London"
  | "Europe/Berlin"
  | "Europe/Moscow"
  | "Africa/Cairo"
  | "Asia/Bangkok"
  | "Asia/Singapore"
  | "Asia/Tokyo"
  | "Australia/Sydney"
  | "Pacific/Auckland";
//#endregion

type CurrentVariables =
  //#region Current weather variables
  | "temperature_2m"
  | "relative_humidity_2m"
  | "apparent_temperature"
  | "is_day"
  | "wind_speed_10m"
  | "wind_direction_10m"
  | "wind_gusts_10m"
  | "snowfall"
  | "showers"
  | "precipitation"
  | "rain"
  | "weather_code"
  | "cloud_cover"
  | "pressure_msl"
  | "surface_pressure";
//#endregion

type DailyVariables =
  //#region Daily variables
  | "weather_code"
  | "temperature_2m_max"
  | "apparent_temperature_max"
  | "temperature_2m_min"
  | "wind_gusts_10m_max"
  | "wind_speed_10m_max"
  | "et0_fao_evapotranspiration"
  | "shortwave_radiation_sum"
  | "wind_direction_10m_dominant"
  | "apparent_temperature_min"
  | "uv_index_clear_sky_max"
  | "uv_index_max"
  | "sunshine_duration"
  | "daylight_duration"
  | "sunset"
  | "sunrise"
  | "precipitation_probability_max"
  | "precipitation_hours"
  | "precipitation_sum"
  | "snowfall_sum"
  | "showers_sum"
  | "rain_sum"
  | "temperature_2m_mean"
  | "apparent_temperature_mean"
  | "cape_mean"
  | "cape_max"
  | "cloud_cover_mean"
  | "cape_min"
  | "cloud_cover_max"
  | "cloud_cover_min"
  | "dew_point_2m_mean"
  | "dew_point_2m_max"
  | "dew_point_2m_min"
  | "wet_bulb_temperature_2m_mean"
  | "wet_bulb_temperature_2m_max"
  | "wet_bulb_temperature_2m_min"
  | "vapour_pressure_deficit_max"
  | "et0_fao_evapotranspiration_sum"
  | "growing_degree_days_base_0_limit_50"
  | "leaf_wetness_probability_mean"
  | "precipitation_probability_mean"
  | "precipitation_probability_min"
  | "relative_humidity_2m_mean"
  | "relative_humidity_2m_max"
  | "relative_humidity_2m_min"
  | "snowfall_water_equivalent_sum"
  | "pressure_msl_mean"
  | "pressure_msl_max"
  | "pressure_msl_min"
  | "surface_pressure_mean"
  | "surface_pressure_max"
  | "surface_pressure_min"
  | "updraft_max"
  | "visibility_min"
  | "visibility_max"
  | "winddirection_10m_dominant"
  | "wind_gusts_10m_mean"
  | "visibility_mean"
  | "wind_speed_10m_mean"
  | "wind_gusts_10m_min"
  | "wind_speed_10m_min";
//#endregion

type HourlyVariables =
  //#region Hourly variables
  | "total_column_integrated_water_vapour"
  | "wet_bulb_temperature_2m"
  | "sunshine_duration"
  | "is_day"
  | "uv_index_clear_sky"
  | "uv_index"
  | "freezing_level_height"
  | "convective_inhibition"
  | "lifted_index"
  | "boundary_layer_height"
  | "cape"
  | "soil_moisture_27_to_81cm"
  | "soil_moisture_9_to_27cm"
  | "soil_moisture_3_to_9cm"
  | "soil_moisture_1_to_3cm"
  | "soil_moisture_0_to_1cm"
  | "soil_temperature_54cm"
  | "soil_temperature_18cm"
  | "soil_temperature_6cm"
  | "soil_temperature_0cm"
  | "temperature_2m"
  | "relative_humidity_2m"
  | "dew_point_2m"
  | "apparent_temperature"
  | "precipitation_probability"
  | "precipitation"
  | "rain"
  | "showers"
  | "snowfall"
  | "snow_depth"
  | "weather_code"
  | "pressure_msl"
  | "surface_pressure"
  | "cloud_cover"
  | "cloud_cover_low"
  | "cloud_cover_mid"
  | "cloud_cover_high"
  | "visibility"
  | "evapotranspiration"
  | "et0_fao_evapotranspiration"
  | "vapour_pressure_deficit"
  | "wind_speed_10m"
  | "wind_speed_80m"
  | "wind_speed_120m"
  | "wind_speed_180m"
  | "wind_direction_10m"
  | "wind_direction_80m"
  | "wind_direction_120m"
  | "wind_direction_180m"
  | "wind_gusts_10m"
  | "temperature_80m"
  | "temperature_120m"
  | "temperature_180m"
  | "shortwave_radiation"
  | "direct_radiation"
  | "diffuse_radiation"
  | "direct_normal_irradiance"
  | "global_tilted_irradiance"
  | "terrestrial_radiation"
  | "shortwave_radiation_instant"
  | "direct_radiation_instant"
  | "diffuse_radiation_instant"
  | "direct_normal_irradiance_instant"
  | "global_tilted_irradiance_instant"
  | "terrestrial_radiation_instant"
  | "temperature_1000hPa"
  | "temperature_975hPa"
  | "temperature_950hPa"
  | "temperature_925hPa"
  | "temperature_900hPa"
  | "temperature_850hPa"
  | "temperature_800hPa"
  | "temperature_700hPa"
  | "temperature_600hPa"
  | "temperature_500hPa"
  | "temperature_400hPa"
  | "temperature_300hPa"
  | "temperature_250hPa"
  | "temperature_200hPa"
  | "temperature_150hPa"
  | "temperature_100hPa"
  | "temperature_70hPa"
  | "temperature_50hPa"
  | "temperature_30hPa"
  | "relative_humidity_1000hPa"
  | "relative_humidity_975hPa"
  | "relative_humidity_925hPa"
  | "relative_humidity_950hPa"
  | "relative_humidity_900hPa"
  | "relative_humidity_850hPa"
  | "relative_humidity_800hPa"
  | "relative_humidity_200hPa"
  | "relative_humidity_250hPa"
  | "relative_humidity_300hPa"
  | "relative_humidity_400hPa"
  | "relative_humidity_500hPa"
  | "relative_humidity_600hPa"
  | "relative_humidity_700hPa"
  | "relative_humidity_150hPa"
  | "relative_humidity_100hPa"
  | "relative_humidity_70hPa"
  | "relative_humidity_50hPa"
  | "relative_humidity_30hPa"
  | "cloud_cover_1000hPa"
  | "cloud_cover_975hPa"
  | "cloud_cover_950hPa"
  | "cloud_cover_925hPa"
  | "cloud_cover_900hPa"
  | "cloud_cover_850hPa"
  | "cloud_cover_800hPa"
  | "cloud_cover_700hPa"
  | "cloud_cover_500hPa"
  | "cloud_cover_600hPa"
  | "cloud_cover_400hPa"
  | "cloud_cover_300hPa"
  | "cloud_cover_250hPa"
  | "cloud_cover_200hPa"
  | "cloud_cover_150hPa"
  | "cloud_cover_100hPa"
  | "cloud_cover_70hPa"
  | "cloud_cover_50hPa"
  | "cloud_cover_30hPa"
  | "wind_speed_1000hPa"
  | "wind_speed_975hPa"
  | "wind_speed_925hPa"
  | "wind_speed_950hPa"
  | "wind_speed_900hPa"
  | "wind_speed_850hPa"
  | "wind_speed_800hPa"
  | "wind_speed_700hPa"
  | "wind_speed_600hPa"
  | "wind_speed_500hPa"
  | "wind_speed_400hPa"
  | "wind_speed_300hPa"
  | "wind_speed_250hPa"
  | "wind_speed_200hPa"
  | "wind_speed_150hPa"
  | "wind_speed_100hPa"
  | "wind_speed_70hPa"
  | "wind_speed_50hPa"
  | "wind_speed_30hPa"
  | "wind_direction_1000hPa"
  | "wind_direction_975hPa"
  | "wind_direction_925hPa"
  | "wind_direction_950hPa"
  | "wind_direction_900hPa"
  | "wind_direction_850hPa"
  | "wind_direction_800hPa"
  | "wind_direction_700hPa"
  | "wind_direction_600hPa"
  | "wind_direction_500hPa"
  | "wind_direction_400hPa"
  | "wind_direction_300hPa"
  | "wind_direction_250hPa"
  | "wind_direction_200hPa"
  | "wind_direction_150hPa"
  | "wind_direction_100hPa"
  | "wind_direction_70hPa"
  | "wind_direction_50hPa"
  | "wind_direction_30hPa"
  | "geopotential_height_1000hPa"
  | "geopotential_height_975hPa"
  | "geopotential_height_950hPa"
  | "geopotential_height_925hPa"
  | "geopotential_height_900hPa"
  | "geopotential_height_850hPa"
  | "geopotential_height_800hPa"
  | "geopotential_height_700hPa"
  | "geopotential_height_600hPa"
  | "geopotential_height_500hPa"
  | "geopotential_height_400hPa"
  | "geopotential_height_300hPa"
  | "geopotential_height_250hPa"
  | "geopotential_height_200hPa"
  | "geopotential_height_30hPa"
  | "geopotential_height_50hPa"
  | "geopotential_height_70hPa"
  | "geopotential_height_100hPa"
  | "geopotential_height_150hPa";
//#endregion

type Minutely15Variables =
  //#region Minutely 15 variables
  | "temperature_2m"
  | "relative_humidity_2m"
  | "dew_point_2m"
  | "apparent_temperature"
  | "precipitation"
  | "wind_gusts_10m"
  | "visibility"
  | "cape"
  | "lightning_potential"
  | "is_day"
  | "shortwave_radiation"
  | "direct_radiation"
  | "diffuse_radiation"
  | "direct_normal_irradiance"
  | "global_tilted_irradiance"
  | "terrestrial_radiation"
  | "terrestrial_radiation_instant"
  | "global_tilted_irradiance_instant"
  | "direct_normal_irradiance_instant"
  | "diffuse_radiation_instant"
  | "direct_radiation_instant"
  | "shortwave_radiation_instant"
  | "sunshine_duration"
  | "freezing_level_height"
  | "snowfall_height"
  | "snowfall"
  | "rain"
  | "weather_code"
  | "wind_speed_10m"
  | "wind_speed_80m"
  | "wind_direction_10m"
  | "wind_direction_80m";
//#endregion

type WeatherModels =
  //#region Weather models
  | "best_match"
  | "ecmwf_ifs"
  | "ecmwf_ifs025"
  | "ecmwf_aifs025_single"
  | "cma_grapes_global"
  | "bom_access_global"
  | "gfs_seamless"
  | "gfs_global"
  | "gfs_hrrr"
  | "ncep_nbm_conus"
  | "ncep_nam_conus"
  | "gfs_graphcast025"
  | "ncep_aigfs025"
  | "jma_seamless"
  | "jma_msm"
  | "jma_gsm"
  | "ncep_hgefs025_ensemble_mean"
  | "kma_seamless"
  | "kma_ldps"
  | "kma_gdps"
  | "icon_seamless"
  | "icon_global"
  | "icon_d2"
  | "gem_global"
  | "gem_regional"
  | "gem_hrdps_continental"
  | "gem_hrdps_west"
  | "meteofrance_seamless"
  | "meteofrance_arpege_world"
  | "meteofrance_arpege_europe"
  | "meteofrance_arome_france"
  | "meteofrance_arome_france_hd"
  | "italia_meteo_arpae_icon_2i"
  | "metno_seamless"
  | "metno_nordic"
  | "icon_eu"
  | "gem_seamless"
  | "knmi_seamless"
  | "knmi_harmonie_arome_europe"
  | "knmi_harmonie_arome_netherlands"
  | "dmi_seamless"
  | "dmi_harmonie_arome_europe"
  | "ukmo_seamless"
  | "ukmo_global_deterministic_10km"
  | "ukmo_uk_deterministic_2km"
  | "meteoswiss_icon_seamless"
  | "meteoswiss_icon_ch1"
  | "meteoswiss_icon_ch2"
  | "geosphere_seamless"
  | "geosphere_arome_austria";
//#endregion

interface WeatherAPIParams {
  latitude: number;
  longitude: number;
  past_days?: number;
  hourly?: HourlyVariables[];
  forecast_hours?: number;
  past_hours?: number;
  temporal_resolution?: TemporalResolution;
  models?: WeatherModels[];
  minutely_15?: Minutely15Variables[];
  forecast_minutely_15?: number;
  past_minutely_15?: number;
  daily?: DailyVariables[];
  tilt?: number;
  azimuth?: number;
  timezone?: Timezone;
  forecast_days?: number;
  wind_speed_unit?: WindSpeedUnit;
  precipitation_unit?: PrecipitationUnit;
  timeformat?: Timeformat;
  temperature_unit?: TemperatureUnit;
  current?: CurrentVariables[];
  cell_selection?: CellSelection;
}